requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
idna==3.3
kucoin-python==1.0.9
python-dotenv==0.19.1
redis==3.5.3
requests==2.26.0
urllib3==1.26.7
websockets==10.0
